Completed photometry figures
Tidied up photometry figures - still need a bit of work.
